Batch updates
Updated batch_creator.py to (hopefully) be fully compatible with the HPC:
- Updated all hpc file paths

- Expanded analysis options functionality to work with Gaia and RUWE tags (still need to add RVs)

- Changed age tag to be more robust (more details later)

- Wrote a test .bash file, which currently shows the Windows path names for testing purposes.

Changed the location of pm and targets_abr (I had previously placed them in the incorrect folder).

Manually added a general approximation for the age of our targets that fits our BHAC models. Done so that the age specification could be done to individual stars rather than to an entire batch.