[WIP] Tried making plotting more general
In the process of creating a function that gets the file names when you input a star. Will make it so that you can pass a star into a function and it does all of the plotting automagically.

Have to get number of companions generated and star mass
Tried to do this from the GUI code, but I hit a lot of roadblocks. My current idea is to open the “all” file and add comment lines containing this information (idea sourced from a StackOverflow thread).