Parallelized ruwe.py!
For now, just parallelized the same way I did with ao.py. There are a few things that I did that might be inefficient.

I also changed some of the lists where we knew the length to be NumPy arrays. I left most of them since I wasn't sure if it would actually be faster to change these.