Fixed ao_filename usage!
There was an issue with my usage of ao_filename and ao_filename[0]: Whenever the program wanted to run ao_filename[0] when no AO file was specified, it would throw an error because ao_filename was being set to [] instead of [''] (even though I told it to!).

I am not sure which change fixed the issue, but this version seems to work! More testing pending...